% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better_hist.R
\name{better_hist}
\alias{better_hist}
\title{A Better Histogram}
\usage{
better_hist(
  .data,
  aes_x,
  label_args = list(),
  add_mean = TRUE,
  add_sd = NULL,
  add_median = NULL,
  add_min = NULL,
  add_max = NULL,
  add_density = FALSE,
  lower_bound = NULL,
  upper_bound = NULL,
  ...
)
}
\arguments{
\item{.data}{dataframe}

\item{aes_x}{aesthestics x argument for `ggplot()`}

\item{label_args}{Arguments for `labs`}

\item{...}{Additional arguments passed to `geom_histogram`}
}
\value{
ggplot2 plot object
}
\description{
An opinionated histogram, using improved binwidth calculations and
embracing an opinionated sense of style.
}
\details{
Calculates binwidth using Freedman-Diaconis method.
}
