% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfr_colors.R
\name{dfr_pal}
\alias{dfr_pal}
\title{Return function to interpolate DFR color palette}
\usage{
dfr_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in dfr_palettes}

\item{reverse}{Boolean indicating whether or not palette should be reversed.}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
Color palette
}
\description{
Return function to interpolate DFR color palette
}
