% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{fmt}
\alias{fmt}
\alias{fmt_comma}
\alias{fmt_compact}
\alias{fmt_pct}
\title{Number Formatting Helpers}
\usage{
fmt_comma(x, decimals = 0, ...)

fmt_compact(x, decimals = 2, drop_trailing_zeros = TRUE, ...)

fmt_pct(x, decimals = 0, ...)
}
\arguments{
\item{x}{Input number}

\item{decimals}{Number of decimal places (default 0)}

\item{...}{
  Arguments passed on to \code{\link[gt:vec_fmt_number]{gt::vec_fmt_number}}
  \describe{
    \item{\code{n_sigfig}}{\emph{Number of significant figures}

\code{scalar<numeric|integer>(val>=1)} // \emph{default:} \code{NULL} (\code{optional})

A option to format numbers to \emph{n} significant figures. By default, this is
\code{NULL} and thus number values will be formatted according to the number of
decimal places set via \code{decimals}. If opting to format according to the
rules of significant figures, \code{n_sigfig} must be a number greater than or
equal to \code{1}. Any values passed to the \code{decimals} and \code{drop_trailing_zeros}
arguments will be ignored.}
    \item{\code{drop_trailing_dec_mark}}{\emph{Drop the trailing decimal mark}

\verb{scalar<logical>} // \emph{default:} \code{TRUE}

A logical value that determines whether decimal marks should always appear
even if there are no decimal digits to display after formatting (e.g., \code{23}
becomes \code{23.} if \code{FALSE}). By default trailing decimal marks are not shown.}
    \item{\code{use_seps}}{\emph{Use digit group separators}

\verb{scalar<logical>} // \emph{default:} \code{TRUE}

An option to use digit group separators. The type of digit group separator
is set by \code{sep_mark} and overridden if a locale ID is provided to \code{locale}.
This setting is \code{TRUE} by default.}
    \item{\code{accounting}}{\emph{Use accounting style}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

An option to use accounting style for values. Normally, negative values
will be shown with a minus sign but using accounting style will instead put
any negative values in parentheses.}
    \item{\code{scale_by}}{\emph{Scale values by a fixed multiplier}

\verb{scalar<numeric|integer>} // \emph{default:} \code{1}

All numeric values will be multiplied by the \code{scale_by} value before
undergoing formatting. Since the \code{default} value is \code{1}, no values will be
changed unless a different multiplier value is supplied. This value will be
ignored if using any of the \code{suffixing} options (i.e., where \code{suffixing} is
not set to \code{FALSE}).}
    \item{\code{suffixing}}{\emph{Specification for large-number suffixing}

\verb{scalar<logical>|vector<character>} // \emph{default:} \code{FALSE}

The \code{suffixing} option allows us to scale and apply suffixes to larger
numbers (e.g., \code{1924000} can be transformed to \verb{1.92M}). This option can
accept a logical value, where \code{FALSE} (the default) will not perform this
transformation and \code{TRUE} will apply thousands (\code{"K"}), millions (\code{"M"}),
billions (\code{"B"}), and trillions (\code{"T"}) suffixes after automatic value
scaling.

We can alternatively provide a character vector that serves as a
specification for which symbols are to used for each of the value ranges.
These preferred symbols will replace the defaults (e.g.,
\code{c("k", "Ml", "Bn", "Tr")} replaces \code{"K"}, \code{"M"}, \code{"B"}, and \code{"T"}).

Including \code{NA} values in the vector will ensure that the particular range
will either not be included in the transformation (e.g.,
\code{c(NA, "M", "B", "T")} won't modify numbers at all in the thousands range)
or the range will inherit a previous suffix (e.g., with
\code{c("K", "M", NA, "T")}, all numbers in the range of millions and billions
will be in terms of millions).

Any use of \code{suffixing} (where it is not set expressly as \code{FALSE}) means
that any value provided to \code{scale_by} will be ignored.

If using \code{system = "ind"} then the default suffix set provided by
\code{suffixing = TRUE} will be the equivalent of \code{c(NA, "L", "Cr")}. This
doesn't apply suffixes to the thousands range, but does express values in
\emph{lakhs} and \emph{crores}.}
    \item{\code{pattern}}{\emph{Specification of the formatting pattern}

\verb{scalar<character>} // \emph{default:} \code{"{x}"}

A formatting pattern that allows for decoration of the formatted value. The
formatted value is represented by the \code{{x}} (which can be used multiple
times, if needed) and all other characters will be interpreted as string
literals.}
    \item{\code{sep_mark}}{\emph{Separator mark for digit grouping}

\verb{scalar<character>} // \emph{default:} \code{","}

The string to use as a separator between groups of digits. For example,
using \code{sep_mark = ","} with a value of \code{1000} would result in a formatted
value of \code{"1,000"}. This argument is ignored if a \code{locale} is supplied
(i.e., is not \code{NULL}).}
    \item{\code{dec_mark}}{\emph{Decimal mark}

\verb{scalar<character>} // \emph{default:} \code{"."}

The string to be used as the decimal mark. For example, using
\code{dec_mark = ","} with the value \code{0.152} would result in a formatted value
of \code{"0,152"}). This argument is ignored if a \code{locale} is supplied (i.e., is
not \code{NULL}).}
    \item{\code{force_sign}}{\emph{Forcing the display of a positive sign}

\verb{scalar<logical>} // \emph{default:} \code{FALSE}

Should the positive sign be shown for positive values (effectively showing
a sign for all values except zero)? If so, use \code{TRUE} for this option. The
default is \code{FALSE}, where only negative numbers will display a minus sign.
This option is disregarded when using accounting notation with
\code{accounting = TRUE}.}
    \item{\code{locale}}{\emph{Locale identifier}

\verb{scalar<character>} // \emph{default:} \code{NULL} (\code{optional})

An optional locale identifier that can be used for formatting values
according the locale's rules. Examples include \code{"en"} for English (United
States) and \code{"fr"} for French (France). We can call \code{\link[gt:info_locales]{info_locales()}} for a
useful reference for all of the locales that are supported.}
    \item{\code{output}}{\emph{Output format}

\verb{singl-kw:[auto|plain|html|latex|rtf|word]} // \emph{default:} \code{"auto"}

The output style of the resulting character vector. This can either be
\code{"auto"} (the default), \code{"plain"}, \code{"html"}, \code{"latex"}, \code{"rtf"}, or
\code{"word"}. In \strong{knitr} rendering (i.e., Quarto or R Markdown), the \code{"auto"}
option will choose the correct \code{output} value}
  }}

\item{drop_trailing_zeros}{Drop zeros after decimal point (default true)}
}
\value{
A character vector
}
\description{
These functions are helpful wrappers around existing formatting functions that
are designed to make it faster and easier to code number formatting.
}
\details{
Included functions:
- `fmt_comma` produces a prettified version of numbers with thousands separator.
- `fmt_compact` produces a compact version of numbers (i.e., 1000 becomes "1K").
- `fmt_pct` produces a percent formatted version of numbers
}
\examples{
# Here's an example numeric vector for these examples:
nums <- c(2032200, 80, 20500000, 984710, 4188.6, 70, 27247000, 9.994, 950, 7e+06,NA)

# fmt_comma is just a pretty print wrapper
fmt_comma(nums)
#>  [1] "2,032,200.00"  "80.00"         "20,500,000.00" "984,710.00"    "4,188.60"
#   [6] "70.00"         "27,247,000.00" "9.99"          "950.00"        "7,000,000.00"
#   [11] "NA"

# fmt_compact produces a compact variant (shortened, with suffixes for million,thousand, etc.)
fmt_compact(nums)
#> [1] "2.03M"   "80"      "20.5M"   "984.71K" "4.19K"   "70"      "27.25M"  "9.99"    "950"
#  [10] "7M"      "NA"

# fmt_pct gives percent sign formatting:
fmt_pct(runif(10))
#> [1] "71.76\%" "52.90\%" "89.98\%" "0.90\%"  "8.16\%"  "55.14\%" "25.21\%" "84.29\%" "29.92\%" "32.05\%"




}
