% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dates.R
\name{fmt_date}
\alias{fmt_date}
\title{Opinionated Date Formatting}
\usage{
fmt_date(
  dates,
  format = c("abbr", "short"),
  range = FALSE,
  range_collapse = NULL
)
}
\arguments{
\item{dates}{vector of dates}

\item{format}{one of "abbr" or "short"}

\item{range}{is this a range or not?}

\item{range_collapse}{strbing for collapse of range, defaults to " to " for
"abbr" format and " - " for "short" format.}
}
\value{
character vector
}
\description{
My own clumsily structured formatting effort for dates
}
\details{
If range is true then it prints a 2-item range as a range.
}
\examples{
x <- as.Date("2025-01-01")
fmt_date(x, "abbr") # Jan 1, 2025
fmt_date(x, "short") # 1/1/2025

a <- c(x, as.Date("2025-07-15"))
fmt_date(a, "abbr", range = TRUE) # Jan 1, 2025 to Jul 15, 2025
fmt_date(a, "short", range=TRUE, range_collapse = " – ") # 1/1/2025 - 7/15/2025

}
