% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_error.R
\name{if_error}
\alias{if_error}
\title{Wrap Function to Return Value If Error Occurs}
\usage{
if_error(.f, .on_error = NA, .verbose = TRUE)
}
\arguments{
\item{.f}{Function to Wrap}

\item{.on_error}{Value to return if error occurs (defaults to `NA`)}

\item{.verbose}{Whether or not to pass on error messages.}
}
\value{
Either function output (on success) or .on_error (if error occurs)
}
\description{
Wrap Function to Return Value If Error Occurs
}
\examples{
log_or_na <- if_error(log) # Returns log() or NA on error
log_or_na(2) # Returns log(2)
log_or_na("A") # Returns NA, but no error.

}
