% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{label_tukey_outliers}
\alias{label_tukey_outliers}
\title{Label Tukey Outlier Values}
\usage{
label_tukey_outliers(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
factor vector
}
\description{
Using tukey outlier fences, labels outlier values
as high, low, extremely high, or extremely low.
}
