% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_with_distribution.R
\name{sample_with_distribution}
\alias{sample_with_distribution}
\title{Creates Sample Using Existing Distribution}
\usage{
sample_with_distribution(x, size, seed = 1234)
}
\arguments{
\item{x}{numeric vector to model distribution from}

\item{size}{A non-negative integer giving the number of items to choose}

\item{seed}{Optional seed for reproducibility}
}
\value{
A numeric vector of size `size` approximating the distribution of `x`
}
\description{
Using an existing vector of numeric values, attempts to
fit a distribution and then generates a random sample
following that distribution.
}
