% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{tukey_fences}
\alias{tukey_fences}
\title{Returns Tukey Outlier Fences for vector}
\usage{
tukey_fences(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
list of upper and lower fences for k=1.5 and k=3
}
\description{
Calculates outlier fences using the Tukey formula:
[Q_1 - k(Q_3 - Q_1) ), ( Q_3 + k(Q_3 - Q_1) ],
}
\details{
with k values for 1.5 (outliers) and 3 (extreme outliers)
}
\examples{
x <- runif(50, 0, 500)
tukey_fences(x)
}
