% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_defined_columns.R
\name{use_defined_columns}
\alias{use_defined_columns}
\title{Use Defined Column Names}
\usage{
use_defined_columns(
  data,
  defined_colnames,
  defined_colname_patterns,
  column_required = c("required", "optional"),
  column_type = c("col_logical", "col_integer", "col_double", "col_character",
    "col_factor", "col_date", "col_datetime", "col_default"),
  keep = TRUE
)
}
\arguments{
\item{data}{Source dataframe}

\item{defined_colnames}{A list of column names to be used in the output.}

\item{defined_colname_patterns}{A list of matching regex patterns to be used
to identify matching column names in the source dataframe.}

\item{column_required}{(optional) whether the defined column must be present
or not in the source dataframe. Defaults to "required" for all defined columns.}

\item{column_type}{(optional) What type each defined column should have.
Defaults to "col_default", which just uses the type present in the source
dataframe for the matching dataframe.}

\item{keep}{Keep non-defined columns in output. Defaults to TRUE.}
}
\value{
.data dataframe with defined column names used.
}
\description{
Processes a dataframe to use pre-defined column names and ensure these names
are present in the dataframe.
}
\details{
This function starts with a source dataframe and a list of desired defined
column names. The dataframe is checked to ensure required defined column names
are present and (optionally) to parse or set specific column data types. The
dataframe is then returned with the defined columns present using the defined
column names instead of the original names, when applicable.

Defined column names are matched to the desired original column names by using
a regex pattern. This allows for matching multiple variations of names in the
source dataframe, provided they have some amount of consistency.

This is useful when designing a function to process or analyze data that may
have variations in naming conventions and eliminates the need to rigidly
enforce a set of naming conventions on end users of the function.

# Defining Column Names
The key to making this function work is using a regex pattern to match source
column names to the desired defined column name. These patterns should ideally
only match a single column name in the source dataframe; if they don't, the
first column name to match will be used in the output (and a warning will be
sent).

# Validating column presence
If a defined column name is marked as "required" (the default), then an error
will be thrown if no matches for this column are found in the source dataframe.

# Column datatypes
Optionally, a list of desired datatypes is provided and the defined columns
are converted into this datatype. Errors will be thrown if parsing into this
datatype fail.

The available specifications are:
- col_logical: logical, contains only TRUE or FALSE
- col_integer: Integers.
- col_double: Doubles.
- col_character: Characters
- col_factor: Fixed set of values.
- col_date: Date type (with day resol  ution)
- col_datetime: POSIXct
- col_default: Keeps the original column type.
}
